deleted extra testfile
